[Fix  #283] 게임 준비 및 시작 로직 위치 변경, Loading 컴포넌트 Error 해결
